# spec/openapi-spec.yml
openapi: 3.0.3
info:
  title: Regular Income
  description: This is the API for Regular Income
  version: 1.0.0
...
paths:
  /api/user/register:
    post:
      tags:
        - UserAuth
      summary: Register new user
      description: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        200:
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/user/{id}/details:
    patch:
      tags:
        - UserAuth
      summary: Updates an user
      description: Updates an  user
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        200:
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/user/{id}:
    delete:
      tags:
        - UserAuth
      summary: Deletes an user
      description: Deletes an  user
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        204:
          description: Successful Operation
  /api/auth/login:
    get:
      tags:
        - UserAuth
      summary: Gets login infos of an user
      description: Gets login infos of an user
      operationId: loginUserInfo
      responses:
        200:
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      tags:
        - UserAuth
      summary: Login an user
      description: Login an user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        200:
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /api/turnover-import/preview/csv:
    post:
      tags:
        - Turnovers
      summary: previews a CSV file to a import preview
      operationId: processCsvPreview
      parameters:
        - in: query
          name: encoding
          required: false
          schema:
            type: string
            default: UTF-8
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: text/csv
                headers:
                  Content-Type:
                    schema:
                      type: string
                      enum: [ text/csv ]
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawCsvTable"
  /api/turnover-import/preview:
    post:
      tags:
        - Turnovers
      summary: Processes a preview to a turnover import
      operationId: processPreview
      parameters:
        - in: query
          name: format
          required: true
          schema:
            $ref: "#/components/schemas/TurnoverImportFormat"
        - in: query
          name: encoding
          required: false
          schema:
            type: string
            default: UTF-8
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: text/csv
                headers:
                  Content-Type:
                    schema:
                      type: string
                      enum: [ text/csv ]
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnoverPreview"
  /api/turnover-import:
    post:
      tags:
        - Turnovers
      summary: Creates a new turnover import
      operationId: createTurnoverImport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                patch:
                  $ref: "#/components/schemas/TurnoverImportPatch"
            encoding:
              file:
                contentType: text/csv
                headers:
                  Content-Type:
                    schema:
                      type: string
                      enum: [ text/csv ]

      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnoverImport"
    get:
      tags:
        - Turnovers
      summary: Get all turnover imports
      operationId: fetchTurnoverImports
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TurnoverImport"
  /api/turnover-import/{id}:
    get:
      tags:
        - Turnovers
      summary: Get a turnover import
      operationId: fetchTurnoverImport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnoverImport"
    delete:
      tags:
        - Turnovers
      summary: Deletes a turnover import
      operationId: deleteTurnoverImport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        204:
          description: Successful operation
    patch:
      tags:
        - Turnovers
      summary: Updates the rows of a turnover import
      operationId: patchTurnoverImport
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurnoverImportRowsPatch"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnoverImport"
  /api/turnover-import/rows:
    patch:
      tags:
        - Turnovers
      summary:
      operationId: batchPatchTurnoverImports
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurnoverImportRowsPatch"
      responses:
        204:
          description: Successful operation
  /api/turnover-import/category/{category-id}:
    get:
      tags:
        - Turnovers
      summary: Get turnovers for a given category
      operationId: fetchTurnoversForCategory
      parameters:
        - in: path
          name: category-id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TurnoverRow"
...
components:
  schemas:
...
    User:
      type: object
      required:
        - id
        - username
      properties:
        id:
          $ref: "#/components/schemas/ID"
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    UserRegistration:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
        firstname:
          type: string
        lastname:
          type: string
...
    RawCsvTable:
      type: object
      required:
        - rows
        - columns
      properties:
        rows:
          type: integer
          format: int32
        columns:
          type: integer
          format: int32
        data:
          type: array
          items:
            type: array
            items:
              type: string
    TurnoverImport:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        importedAt:
          $ref: "#/components/schemas/Timestamp"
        filename:
          type: string
        filesizeBytes:
          $ref: "#/components/schemas/Filesize"
        format:
          $ref: "#/components/schemas/TurnoverImportFormat"
        encoding:
          type: string
        firstTurnover:
          $ref: "#/components/schemas/Date"
        lastTurnover:
          $ref: "#/components/schemas/Date"
        turnovers:
          type: array
          items:
            $ref: "#/components/schemas/TurnoverRow"
    TurnoverImportPatch:
      type: object
      properties:
        format:
          $ref: "#/components/schemas/TurnoverImportFormat"
        encoding:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TurnoverRow"
    TurnoverImportRowsPatch:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TurnoverRowPatch"
        deleteRowIds:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ID"
    TurnoverRowPatch:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        categoryId:
          $ref: "#/components/schemas/ID"
    TurnoverPreview:
      type: object
      properties:
        filename:
          type: string
        format:
          $ref: "#/components/schemas/TurnoverImportFormat"
        encoding:
          type: string
        uploadTime:
          $ref: "#/components/schemas/Timestamp"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TurnoverRowPreview"
    TurnoverRowPreview:
      type: object
      allOf:
        - $ref: "#/components/schemas/TurnoverRow"
        - properties:
            importable:
              type: boolean
            suggestedCategories:
              type: array
              items:
                $ref: "#/components/schemas/CategorySuggestion"
    CategorySuggestion:
      type: object
      required:
        - categoryId
        - precedence
        - frequency
      properties:
        categoryId:
          $ref: "#/components/schemas/ID"
        precedence:
          type: number
          format: double
        frequency:
          type: integer
          format: int32
        origins:
          type: array
          items:
            type: string
    TurnoverRow:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        date:
          $ref: "#/components/schemas/Date"
        amountInCents:
          $ref: "#/components/schemas/MoneyValue"
        categoryId:
          $ref: "#/components/schemas/ID"
        checksum:
          type: string
        similarityChecksum:
          type: string
        description:
          type: string
        suggestedCategory:
          type: string
        recipient:
          type: string
    TurnoverImportFormat:
      type: string
      enum:
        - VR_BANK_CSV
        - DKB
        - NEW_DKB
        - PAYPAL
...
    ID:
      type: string
      format: uuid
    MoneyValue:
      type: integer
      format: int32
    Filesize:
      type: integer
      format: int64
    Date:
      type: string
      format: date
    Timestamp:
      type: string
      format: date-time
...